Only compile the regular expression once
There is more benefit to compiling a regular expression if you only have
to compile it once. Moving it from inside a function to the module
level reduces the number of times it's compiled from once per time
`parse` is called to once per interpreter instantiation. To achieve
this, the code inside `_get_duration_re` is being moved to the module
level and the function is being removed.

One implementation change of note: a list comprehension is being used to
build the list of expressions to compile so that its memory is allocated
only once.